backend/app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
backoff==2.2.1
bcrypt==4.1.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.2.6
langchain-openai==1.1.7
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.2
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
mem0ai==1.0.1
numpy==2.4.0
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
passlib==1.7.4
portalocker==3.2.0
posthog==7.5.0
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.40.0
xxhash==3.6.0
zstandard==0.25.0
